Hack to work around xf86ModeStatusToString removal from public headers
src/via_display.c (xf86ModeStatusToString): HACK: Copy the prototype for
this function from xf86_priv.h. It is used in a log message in
via_crtc_mode_fixup. This makes it compile for now.

TODO: In the long run, either this function should be moved back to the
public headers in xserver, or the driver should stop using it (but then
the log message will be less informative).