configure.ac: Disable XVMC support, does not compile (#6)
Comment out the lines enabling XVMC for x86 architectures. It currently
does not compile. It depends on at least one private header
("xf86xvpriv.h") that is not installed by XLibre. It turns out that
Fedora has not been compiling this part of the driver since 2021, by
simply not installing the build dependencies for it.

Fixes #6.